refactor(actions): extract the branch card's action-runner machinery into shared modules
BranchCardActionsBar owned the entire action-running experience in one
~1,040-line component, so the repo card (whose Run button is still a
stub on top of run_repo_action) could only get the same behavior by
duplicating it. Pull the reusable pieces out into the actions feature,
parameterized per the plan note by exactly three things — an opaque
scope id (branch id or synthetic repo scope id), a loadActions()
callback, and a run(actionId) callback:

- actionRunner.svelte.ts: the ActionRunner state machine. Owns the
  configured action list, the live running-execution set (hydrated via
  get_running_branch_actions, updated by the action_status and
  run-phase event subscriptions), stopping/fade-out bookkeeping, and
  the output modal state, plus the run / stop / show-output /
  run-again handlers and the grouped-action deriveds.
- RunningActionPills.svelte and PrimaryRunActionButton.svelte: the
  secondary running-pill row and the primary run button (circular or
  endpoint pill with copy-URL), moved verbatim with their styles and
  transitions, driven by a runner prop. The endpoint URL rewriting
  stays branch-side and comes in via a getEndpointCopyUrl prop; a show
  prop keeps the setting-up gate inside the same if-block so the slide
  transition behavior is unchanged.
- actionMenu.ts + ActionsSubmenu.svelte: the "Actions" submenu builder
  (menu item types, per-type icons, the Format & Check collapsing) as
  a pure function, and the DropdownMenu.Sub that renders it.
- actionGroups.ts: the five pure grouping helpers move over from
  branchCardHelpers.ts (BranchCardActionsBar was their only consumer).
- altKey.svelte.ts: shared refcounted Alt-key tracking for the
  alt-click-to-stop affordance, previously component-local listeners.

BranchCardActionsBar becomes the first consumer — scoped to branch.id,
listProjectActions, and runBranchAction — with zero behavior change,
keeping only its branch-specific concerns (endpoint rewriting, Open In
menu, rename/rebase/squash/delete items, the output modal's branch id).
RepoCard is deliberately untouched; a follow-up session wires it as the
second consumer on top of run_repo_action.

Verified with `just typecheck`, `just fmt-check`, `just test-frontend`
(481 passed), and a `VITE_REPOS_UI_ENABLED=true` production build.
